requirements.txt
torch>=1.8.0
torchvision
transformers
tqdm
numpy
pandas
scikit-learn
Pillow
pyyaml 